user-setup: support abs/rel paths for cert options
Support both absolute and relative paths for the cert options, instead
of relative paths only, by using realpath(1) to resolve the arguments.

(Note that the script cannot reference the path arguments verbatim
because much of the script operates in the workspace's toplevel
directory, which is a different location than where the script is
invoked. In that situation, relative paths would not resolve.)

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>