install-debuginfo-local: fix installation with new lookaside locations
Debuginfo packages now install binary debuginfo to /usr/share/debug and
source debuginfo to /usr/share/debug/src.

The purpose for the migration is for these new locations to serve as
"lookaside" directories used by future versions of clr-debug-info,
enabling select debuginfo packages to be installed in bundles and safely
coexist with the FUSE mounts.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>