Support cloning of pre-existing package repos
Over the lifetime of Clear Linux OS, many packages have been added and
removed, but the package *repos* always remain once they have been
created.

Thus, if a package is removed at one point, and later added back to the
distro, the same package repo should be used. This commit adds support
to `make autospecnew` for that use case, always using a pre-existing
package repo if one is available.

(Note that the gitolite repo detection is more complicated because it
must support configurations with and without "wild repos" enabled.)

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>