Makefile.common: Add bump.list to make koji
Add the ability for make koji to look a bump.list file in the package
directory and make and koji all the packages listed in that file.

This is useful for package's that depend on the updated package and
should be rebuilt after the original package is.

If for some reason this functionality isn't desired make koji-nowait
will not look at the bump.list.

Signed-off-by: William Douglas <william.douglas@intel.com>